updates notification styles
