+++ app/containers/Notifications/Notification.scss
@@ -11,5 +11,8 @@
 
 .icon svg path{
   fill: white;
-  opacity: 0.75;
+}
+
+.warning {
+  opacity: 0.75
 }
