+++ app/containers/Notifications/Notifications.jsx
@@ -3,6 +3,7 @@ import React, { Component } from 'react'
 import { isEqual, difference } from 'lodash'
 import ReactNotificationSystem from 'react-notification-system'
 import style from './Notification.scss'
+import overrideStyles from './overrideStyles'
 
 import WarningIcon from '../../assets/icons/warning.svg'
 import CheckIcon from '../../assets/icons/check.svg'
@@ -10,61 +11,6 @@ import CheckIcon from '../../assets/icons/check.svg'
 type Props = {
   notifications: Array<NotificationType>,
   hideNotification: Function
-}
-
-const defaultWidth = 480
-const marginLeft = -(defaultWidth / 2)
-
-const overrideStyles = {
-  Containers: {
-    DefaultStyle: {
-      width: defaultWidth
-    },
-    tc: {
-      marginLeft
-    },
-    bc: {
-      marginLeft
-    }
-  },
-
-  NotificationItem: {
-    DefaultStyle: {
-      borderTop: 'none',
-      height: '60px',
-      marginTop: 0,
-      display: 'flex',
-      alignItems: 'center',
-      paddingLeft: '60px',
-      color: '#FFF'
-    },
-
-    success: {
-      backgroundColor: '#5ABF6B'
-    },
-    error: {
-      backgroundColor: '#EE6D66'
-    }
-  },
-
-  Title: {
-    DefaultStyle: {
-      display: 'none'
-    }
-  },
-
-  Dismiss: {
-    DefaultStyle: {
-      top: '22.5px',
-      right: '25px',
-      height: '22px',
-      width: '22px',
-      fontSize: '22px',
-      fontWeight: 500,
-      backgroundColor: 'none',
-      opacity: '0.75'
-    }
-  }
 }
 
 class Notifications extends Component<Props> {
@@ -91,7 +37,11 @@ class Notifications extends Component<Props> {
           this.rnsRef.addNotification({
             children: (
               <div className={style.icon}>
-                <WarningIcon />
+                {notifications[0].level === 'success' ? (
+                  <CheckIcon />
+                ) : (
+                  <WarningIcon className={style.warning} />
+                )}
               </div>
             ),
             uid: notificationId,
